yolo-service/requirements.txt
fastapi==0.111.0
uvicorn==0.29.0
ultralytics==8.2.19
pillow==10.3.0
python-multipart==0.0.9
torch
torchvision
torchaudio
